+++ contracts/common/interfaces/IBookOfDocuments.sol
@@ -55,7 +55,7 @@ interface IBookOfDocuments {
 
     function isSubmitted(address body) external view returns (bool);
 
-    function qtyOfDocuments() external view returns (uint256);
+    function qtyOfDocuments() external view returns (uint);
 
     function docs() external view returns (bytes32[]);
